+++ core/src/main/scala/replpp/DottyReplDriver.scala
@@ -1,42 +1,41 @@
-package dotty.tools.repl
+package replpp
 
 import scala.language.unsafeNulls
-
-import java.io.{File => JFile, PrintStream}
+import java.io.{PrintStream, File as JFile}
 import java.nio.charset.StandardCharsets
-
-import dotty.tools.dotc.ast.Trees._
+import dotty.tools.dotc.ast.Trees.*
 import dotty.tools.dotc.ast.{tpd, untpd}
 import dotty.tools.dotc.config.CommandLineParser.tokenize
 import dotty.tools.dotc.config.Properties.{javaVersion, javaVmName, simpleVersionString}
-import dotty.tools.dotc.core.Contexts._
-import dotty.tools.dotc.core.Decorators._
-import dotty.tools.dotc.core.Phases.{unfusedPhases, typerPhase}
+import dotty.tools.dotc.core.Contexts.*
+import dotty.tools.dotc.core.Decorators.*
+import dotty.tools.dotc.core.Phases.{typerPhase, unfusedPhases}
 import dotty.tools.dotc.core.Denotations.Denotation
-import dotty.tools.dotc.core.Flags._
+import dotty.tools.dotc.core.Flags.*
 import dotty.tools.dotc.core.Mode
 import dotty.tools.dotc.core.NameKinds.SimpleNameKind
 import dotty.tools.dotc.core.NameKinds.DefaultGetterName
-import dotty.tools.dotc.core.NameOps._
+import dotty.tools.dotc.core.NameOps.*
 import dotty.tools.dotc.core.Names.Name
-import dotty.tools.dotc.core.StdNames._
+import dotty.tools.dotc.core.StdNames.*
 import dotty.tools.dotc.core.Symbols.{Symbol, defn}
 import dotty.tools.dotc.interfaces
 import dotty.tools.dotc.interactive.Completion
 import dotty.tools.dotc.printing.SyntaxHighlighting
-import dotty.tools.dotc.reporting.{ConsoleReporter, StoreReporter}
-import dotty.tools.dotc.reporting.Diagnostic
+import dotty.tools.dotc.reporting.{ConsoleReporter, Diagnostic, StoreReporter, UniqueMessagePositions}
 import dotty.tools.dotc.util.Spans.Span
 import dotty.tools.dotc.util.{SourceFile, SourcePosition}
 import dotty.tools.dotc.{CompilationUnit, Driver}
 import dotty.tools.dotc.config.CompilerCommand
-import dotty.tools.io._
+import dotty.tools.io.*
+import dotty.tools.repl.*
 import dotty.tools.runner.ScalaClassLoader.*
-import org.jline.reader._
+import org.jline.reader.*
+import DottyRandomStuff.newStoreReporter
 
 import scala.annotation.tailrec
 import scala.collection.mutable
-import scala.jdk.CollectionConverters._
+import scala.jdk.CollectionConverters.*
 import scala.util.control.NonFatal
 import scala.util.Using
 
@@ -219,7 +218,7 @@ class DottyReplDriver(settings: Array[String],
     }
 
     if expr.startsWith(":") then
-      ParseResult.commands.collect {
+      DottyRandomStuff.ParseResult.commands.collect {
         case command if command._1.startsWith(expr) => makeCandidate(command._1)
       }
     else
